backend/roadpack-build/requirements.txt
psycopg[binary]>=3.1.18
httpx>=0.27.0
python-dotenv>=1.0.0
shapely>=2.0.0
